cairo: reduce cost of measuring bounding boxes
Measuring the size of the stroke and fill bounding boxes is _very_ expensive
(the flattened path is tesselated) when rendering zoomed in SVGs, this due to
cairo subdividing until the tolerance value of cairo is met. This commit
momentarily drops that precision to a precision of 1px during bounding box
measurements.

If these computations could be avoided, using the bounds of the path would be
desirable. Some optimizations to further reduce edge case costs in cairo are to
be expected.