Avoid unnecessary strdups
In all cases except one, the property bag does not need to store copies
of the attributes and values. So make the property bag store them by
reference, and add rsvg_property_bag_dup for the one case where we do
need to make a bag that owns the attributes and values.