Fix rsvg.h causing an "internal" deprated warning
Before this patch rsvg.h marks the RsvgSizeFunc *type* as deprecated, and
then uses it to declare other deprecated functions, causing gcc to always
emit a depecrated warning as soon as rsvg.h is included. This breaks
compilation of apps which compile with -Werror.

Fix this by not making the RsvgSizeFunc type deprecated, but leaving
the function prototypes using it as deprecated.

https://bugzilla.gnome.org/show_bug.cgi?id=665905