bgo#744299 - Ensure the type of pattern fallbacks
Atte Kettunen's fuzz testing yielded an SVG with a pattern paint server that
had an xlink:href to a *rect*, not to another patern.  Since we were not checking
type type of resolved nodes when applying pattern fallbacks, we were using a
structure of the wrong type.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=744299

Signed-off-by: Federico Mena Quintero <federico@gnome.org>