Internally use 'char *' for string data
There was a mixture of guint8*, guchar*, gchar*, etc.  Use char* because that's
about as sane as we can get, and only cast where we can't change the public API
or where we must interface with other libraries.  *Cough* xmlChar *cough*.

This removes a metric assload of compiler warnings, and I'm fine with that.