rsvg_path_arc(): take an RsvgPathBuilder, not a full RSVGParsePathCtx
The ctx was used to extract the start point for the arc, and to
set the 'current point' to the arc's end point when the drawing
is finished.

We now pass in extra arguments for the start point (which don't come
with the path builder).

Also, added whitespace in the math-y part, because without whitespace
it makes me sad.

Signed-off-by: Federico Mena Quintero <federico@gnome.org>