bgo#739329 - Remove single quotes around font-family names before passing them to Pango
If you have an SVG with

  <text style="font-family: 'New Century Schoolbook';"> ... </text>

our CSS-parsing code simply passes the singly-quoted string to Pango.  In turn,
Pango doesn't like the quotes because they are not part of the font's name.

This is a simple patch to remove single quotes from *all* style attributes, not
just strings (which is where the CSS spec allows them).  We need a real CSS
parser, not something built out of g_strsplit().