rsvg-cairo-draw: Take 'invisible' objects into account for bounding boxes
Bounding boxes were computed only for objects that have a fill or a stroke.  However,
for symbolic icons (e.g. adwaita-icon-theme/src/symbolic/gnome-stencils.svg) we can have
fill-less and stroke-less rectangles that are just used to center the actual icons on
the area they are allocated.  In that case we need to compute the full bounding box
for the allocated area, i.e. the fill-less/stroke-less rectangles.

This should not affect 'normal' SVGs where all objects have a fill or a stroke.