Windows Builds: Add a Pre-Configured config.h Template
This adds a config.h.win32.in template that is processed during autogen/
configure, in which the processed config.h.win32 can be later dist'ed so
that Visual Studio project files can make use of this for building librsvg,
since Visual Studio builds by themselves do not use autotools.

Also add in the autotools build files so that this is processed, and this
and the processed config.h.win32 is dist'ed.

https://bugzilla.gnome.org/show_bug.cgi?id=753555