Build: Add MSVC Project Files
Allow librsvg to be built directly with Visual Studio via the use of
project files, like what is done now for GTK+ etc.  Note that the
projects for the main librsvg DLL and rsvg-convert, as well as the
property sheets to copy the build results and headers are generated from
templates, which is done during 'make dist', to ease future maintenance,
which is done via the common autotools module.

The other common autotools module is then used to create the 2012/2013/
2015 projects from the 2010 projects, as they are largely similar in
format.

https://bugzilla.gnome.org/show_bug.cgi?id=753555