build: Add common autotools module for introspection under MSVC
This adds a common autotools module that is used to generate the
full command line for generating the .gir files for librsvg and
under Visual Studio's NMake, as well as the file lists of files that
are to be passed into the introspection scanner.  This is exactly the
same file (as listed in the file's comment block) found under the
same name in $(gi_src_root)/build, so updates for this comes
from there.

https://bugzilla.gnome.org/show_bug.cgi?id=765245