Visual Studio builds: Fix pre-Visual Studio 2012 builds
A recent commit made use of the C99 function isnan(), which is not
supported directly in pre-2013 Visual Studio.  Earlier Visual Studio
however does provide _isnan() which does what we will need, so add a
fallback math.h in build/win32/vs9 that defines isnan() and includes the
stock math.h, and make the projects include this math.h.

This header is copied to the build/win32/vs[10|11] during 'make dist', as
Visual Studio 2010 and 2012 will need this as well.