bgo#764808 - start off with state->current_color = black
When SVG comes in with e.g. fill="currentColor", it means that the content is likely
being embedded into other content, and expects the CSS to cascade appropriately.  For
example, if there is an SVG for an equation and it gets inserted into an HTML paragraph
of text, then the equation should have the same color as the text.

However, librsvg doesn't allow passing in a current-state-of-CSS; it always starts
off fresh.  So, here we make it start with the current_color being opaque black,
instead of transparent (all-bits-zero due to initialization).

This will at least make things visible.  Later we can explore inheriting the cairo_t's
current colors.

https://bugzilla.gnome.org/show_bug.cgi?id=764808