Start a branch to port bits of librsvg to Rust
I want parts of librsvg to be rewritten in Rust so that the library's
internals are written in a safe language.  It should be possible to
preserve the public C API, while internally parts of librsvg are
written in Rust.

I don't know how to integrate the Rust-built library into the C one yet,
so this is a work in progress.

To begin, we have path_to_segments() from rsvg-marker.c.  The rest of
the marker-drawing machinery is still missing; and the ported version
of that function needs some tests - it doesn't do the right thing with a
closepath operator.