Rust: Remove the need for segment_num
`segment_num` only kept track of the index of the last element in the
`segments` vector, but the vector already keeps track of that itself.
Using its `last_mut()` method instead, we can get rid of the counter
and its associated logic, and also get to use `expect()` with an
appropriate error message, instead of getting a plain "index out of
bounds" error when there is no segment yet.