marker.rs: Handle Cairo's artificial moveto after a closepath
Cairo outputs a moveto after every closepath, so that subsequent
lineto/curveto commands will start at the closed vertex.

We don't want to generate a point (a degenerate segment) for that
artificial moveto.

To fix this, we add two states, State::Initial and State::ClosedSubpath,
so that we can differentiate between the very first moveto in a sequence
of path commands, and the state where we just closed a subpath and we
expect an incoming artificial moveto.