rsvg-base: New function rsvg_acquire_node_of_type()
In many places where we call rsvg_acquire_node(), the next thing
done is to check that the type of the acquired node is the one
that particular code expects.  The node is released if it doesn't match.

Encapsulate that pattern in a function so we can use
it everywhere instead of checking by hand.

Also, we allow rsvg_node_acquire() to take a NULL uri, for convenience.