marker.rs: Port rsvg_rust_render_markers() to Rust
This is the interesting part of the old rsvg_render_markers() from C.

That function in C now remains only as a wrapper that extracts a few
values from the RsvgDrawingCtx and passes them on to
rsvg_rust_render_markers().  I'll probably add accessor functions later
so that this kind of wrapper is not necessary.

Also, change the Cargo.toml to build a staticlib.  Re-enable the Rust
build in the master Makefile.am.

And it works!  The tests pass!

The Makefile.am is fishy; it uses rust/target/debug/*.a instead of
automatically switching between "debug" and "release".  Also, this is
not yet integrated into configure.ac at all.

But it works!