tests: Add a README.md; make rsvg-test.c's test names more clear
We have a new README.md that says how the tests work and how to deal
with failed tests.

Additionally, the GTest test names that rsvg-test produces are now
easier to understand.  The test name

  /rsvg-test/reftests/foo/bar.svg

will actually match the file

  fixtures/reftests/foo/bar.svg

Whereas before, the test for that file would be called
"/rsvg/reftest/foo/bar.svg", which isn't entirely obvious.