RsvgLength: replace the 'char factor' field with a 'unit' enum
We now have a LengthUnit enum for use with RsvgLength, instead of representing
units with simple chars.

Along with this change, the filters code was hackily sticking an 'n' into some_length.factor
to really mean, 'this length didn't get specified in the SVG data'.  Replaced that
with boolean foo_specified fields that the code later checks to see if the
length in question is available.

The same pattern was used in the Text code; also made it use foo_specified booleans.