Move rsvg_render_markers() completely into Rust
We used to have an rsvg_render_markers() in C, and after pulling some
data out of ctx and state, it would call the actual implementation in
Rust.  Now that we can extract that data from ctx with accessor
functions, we can have the whole implementation of rsvg_render_markers()
in Rust.