rsvg-styles: Use g_message() instead of g_warning() for errors from libcroco
Our test suite considers a test to fail if it calls g_warning().  Our
code that uses libcroco would call g_warning() on invalid or unsupported
CSS in <style> blocks, and this would causes tests to fail as if there
were something wrong with out code, not with the CSS itself.

We'll have to have a way to report CSS errors upstream, and tests for
specific kinds of errors.