Radial gradients: fix the focus point so it lies within the gradient's circle
An SVG radial gradient is from a focus point (fx, fy) towards a
circle (cx, cy, r).  The focus point is assumed to be inside the circle.

If the focus point is specified to be outside the circle, we need to
find the closest point to it in the circumference, i.e. the point in the
circumference that intersects the line that passes through (cx, cy)
and (fx, fy).

This computation makes us pass the reference test - regenerated the
corresponding image.