RsvgNode: don't keep the node's element name in this struct while parsing
While an RsvgHandle is parsing, it keeps a stack with the name of the
current XML element.  However, it kept the element name as part of the
RsvgNode structure.  The name pointers are only valid while the
xmlParserCtxtPtr is alive; after that, node->name contains a dangling
pointer (fortunately unused).

Instead of burdening RsvgNode with maintaining the parsing stack, we now
keep this stack in RsvgHandlePrivate instead.