NodeCreator: Add a supports_class_attribute field, and populate it from the spec
https://www.w3.org/TR/SVG/attindex.html#RegularAttributes has a list of
which elements support the "class" attribute for CSS styling.

Right now most of our node implementations fetch the "id" and "class" by
hand, and feed them to rsvg_parse_style_attrs().  We'll factor out that
call for all the node types, and do the call in a single place in the
toplevel node_set_atts().

According to the spec all elements support "id".  The NodeCreator
structure, and the node_creators[] array of supported element types,
encode which elements support the "class" attribute, so we can fetch it
as appropriate.