gradient.rs: Start an implementation of gradient parameters in Rust
Gradients can inherit their attributes from "fallback" gradients
mentioned by IRI.  The C code does a little magic to resolve the
fallbacks: it does a shallow-copy of an RsvgLinearGradient or
RsvgRadialGradient structure, and then walks the chain of fallbacks,
filling in the shallow-copied structure as appropriate.  This is done
even for the children of the gradient node, i.e. its color stops: the
pointer to the node->children is shallow-copied (!)

For the Rust version, we'll have specific structures to hold a
gradient's semantics.  They'll have Option<> fields that can be resolved
from fallbacks.  We will have a "digested" array of color stops, not a
high-level array of child nodes that may or may not actually contain
stop elements.