gradient.rs: Avoid copying a gradient when resolving
We used to call into the C helper code to get a Gradient from a URI.
Then we would clone() that gradient and operate on it.

Now we operate directly on the Gradient that we got from the C helper
code, without cloning it.  We create a box from that Gradient pointer
and pass that Box around.  We own that Gradient, after all!