Visual Studio builds: Add NMake Makefile for the Rust bits
We need to call cargo to build the rust portions of librsvg, before we can
link the final librsvg DLL, so add an NMake Makefile that can be used to
build and clean the parts, with the needed envvars set.

This will indicate that we will no longer support Visual Studio 2012 (or
earlier) builds, as Rust will generate code that will need items in the
Visual Studio 2013 (and later) CRTs during linking.

More commits to follow in this regard, but this paves the first step to
support Visual Studio builds for the rust-ified librsvg.