MSVC builds: Make MSVC 2013 the baseline MSVC Version
Due to the rustification of librsvg, it is now only possible to support Visual
Studio 2013 and later as the Rust compiler will require items from the CRT
during linking that is provided by the MSVC 2013 (and later) CRTs.

As a result, drop the Visual Studio 2008~2012 projects, and we no longer need
a fallback math.h implementation since these are now covered by the CRT
implementation.