Use a hard-coded font file to render test images, for predictable font rendering
If we depend on whatever Fontconfig finds on the system (or the user's
configuration, or any number of variables not under our control), "make
distcheck" will certainly fail.

Now we ship a specific font file (LiberationSans-Regular.ttf) that is
used only for the reftests.  We use that in addition to explicitly
setting font rendering options for tests.

Also, re-normalize all the reference images to this new font.

Thanks to Vlad Vukicevic for the idea to use a hard-coded font file, and
to Christian Hergert for the Fontconfig / Pango incantations.