cnode.rs: New stub for having a node implemented in C, accessible from Rust
node.rs: The node_impl needs to be a Box<NodeTrait>, as the actual
implementation will live in the heap:  can't have a lifetime specifier
there.

We also have a new rsvg_rust_cnode_new() to create nodes from C and wrap
the C structures that will actually implement them.