Allow pattern specifications to have a fallback color, per the spec
See here: https://www.w3.org/TR/SVG/painting.html#SpecifyingPaint

A paint specification can be a color, or "<funciri> [somecolor]".  If
the optional color is specified, it will be used in case the paint
server specified by the IRI is not valid (not found, doesn't have enough
parameters, or somehow doesn't manage to paint anything).