Test pservers-pattern-09-f.svg from the official test suite
I believe the official reference image is wrong.  The test SVG says that
both cases in the file should render the fallback color (green), and
that is indeed what we do.  However, the official reference image only
renders the first case in green, and doesn't render the second case at
all.