gradient.rs, pattern.rs: Do the fallbacks with a macro for less verbosity
Instead of a bunch of

  if self.foo.is_none () { self.foo = Some (default_value); }

for defaults, or instead of

  if self.foo.is_none () { self.foo = fallback.foo; }

for fallbacks, we now have a macro that can be invoked as

  fallback_to! (self.foo, some_value);

which seems easier to read.

The macro is duplicated for both gradient.rs and pattern.rs, but we will
refactor those two files into a single thing anyway.