Visual Studio builds: Support Visual Studio 2017
Update the autotools scripts so that we can support Visual Studio 2017 by
copying the 2013 projects and update items as necessary to produce the
2017 projects

Note that the toolset version string format changed for Visual Studio 2017, so allow
one to use a custom toolset version string, otherwise a default toolset version string
string will be generated as it was before.

Note also that Visual Studio 2017 aims to be compatible with Visual Studio 2015 on the
CRT level, so it should be possible to use 2017-built binaries with 2015-built binaries.