Visual Studio builds: Move projects to win32/
Move the projects to win32/ from build/win32/, so that one will need to go down
one less level down the tree to reach the project files, and will allow the
autotools modules (Makefile.msvcproj, Makefile-newvs.am,
Makefile.msvc-introspection) to be in sync with the ones in GLib and G-I master.

This also makes the support of Visual Studio 2017 complete by allowing it in the
NMake Makefiles, which is a must for this package since NMake Makefiles are used
to build the Rust bits on Visual Studio, as well as for introspection builds.