node.rs: Add a Node::with_impl() method, to access the NodeTrait impl out of band
This is a bit sketchy, because we use the downcast-rs crate to actually
be able to downcast a our "node_impl: Box<NodeTrait>" field into the
concrete implementation type.  At first we'll use this in the marker.rs
implementation, as markers are not rendered from the normal ::draw()
method.  Maybe we'll find a way to make this cleaner at some point.