Backport: bgo#777834: Don't try to render text whose PangoLayout has zero extents
From commit 8bf4354fdc3703c0ea6ecec86ce2f3583933c03a

Text elements sometimes have children with empty strings for their
contents, so PangoLayout returns an ink_rect with (width, height) = (0, 0).

In this case, we would set a bbox's size to (0, 0), then use that in an
affine matrix, and try to invert it, and kaboom.

https://bugzilla.gnome.org/show_bug.cgi?id=777834