Visual Studio builds: Support Visual Studio 2017
This adds Visual Studio 2017 support for the 2.40.x series by updating the
autotools scripts that copy the 2010 projects and updating the items in
there as needed.  Since Visual Studio 2017 has a different toolset version
string format, a custom toolset version string is allowed and used if one
is specified, otherwise it will be generated as it is before.

Since Visual Studio 2017 aims to be compatible with Visual Studio 2015 on
the CRT level, one should be able to use 2017-built binaries with
2015-built binaries.