node.rs: Start defining node::Error type
This in the context of https://bugzilla.gnome.org/show_bug.cgi?id=776932

A "transform (0, 0, 0, 0, 0, 0)" attribute ripples down the code, so
that we try to invert that matrix and panic.

We'll keep a "result" field inside struct Node.  When a node
parses/validates its attributes, it will set that result to Err if it
finds a problem.  This is how we'll remember that a node "is in error",
to use the SVG spec's terminology.

Later, Node::draw() will not render a node that is in error.  We can
also do optimizations in the call stack above Node::draw(); for example,
making rsvg_node_draw_from_stack() ignore nodes that are in error so
that it can avoid doing a state inheritance for a node in error.