property_bag.rs: New function parse_or_default()
This looks up a key in the property bag.  If the key is not found,
returns a default value.

If the key is found, then the string value is parsed.  If the parse is
successful, that value is returned.  Otherwise, we turn that into a
NodeError with a parse error inside.

Make marker.rs use that for the "aspect" attribute.