NodeSvg: Only access self.atts if it is non-null
The contents of self.atts will be null if an <svg> element has no
attributes.  This is because node_set_atts() in the C code will not call
a Node.set_atts() method if the property bag is empty.

While that is an optimization to avoid a bunch of hash lookups on empty
hash tables, it does mean that node implementations cannot assume that
their set_atts() method is always called.  I'm not sure that I like
this.  Maybe we should just profile things with/without the extra hash
lookups and see if they make a difference.