Add struct-use-04-b.svg from the SVG 1.1 test suite
This file modified to look for includes in the "resources" subdirectory.

Also, rename the "images" directory to "resources", and modify
rsvg-test.c to ignore that directory as well.  This is because
struct-use-04-b.svg includes another SVG file inside resources/, and if
that directory weren't ignored, rsvg-test would look inside it for SVG
files to try rendering as tests.