Backport: bgo#634324 - Gaussian blur with negative-scaling transformation matrix is not rendered
From commit f7ccfbff857b5ec8d7ace494e20f28d97f1c91f1

When we figure out the standard deviations for Gaussian blurs, we were
simply multiplying the given values by the scaling factors in the current
transformation.  But if this is a negative-scaling matrix, we'd get
negative values, and we would not do a blur at all.

Take the absolute value of the scaled factors instead.

https://bugzilla.gnome.org/show_bug.cgi?id=634324