pattern.rs: Fully implement NodePattern in Rust. Yay!
We had an auxiliary Pattern structure, which is what the patterns
machinery used to resolve patterns at rendering time.  However, the SVG
data for patterns was still being kept in the C code.  Now we have a
Rust NodePattern, as expected, and the C code does not have any patterns
machinery anymore (other than RsvgPaintServer's to actually call up the
pattern.)