win32/rsvg-rust.mak: Streamline Rust build process
As the recommended way to setup Rust is via rustup, set the default
toolchain/target during the NMake build process, so that one will only
need to have the Rust executables in their PATH and ensure that they
have the needed toolchains (i686-pc-windows-msvc and/or
x86_64-pc-windows-msvc) installed via rustup.

Please note that we are being conservative here-we use the stable
releases of the toolchains only.  If using a beta/nightly toolchain
release, one needs to update this file for this purpose.