feGaussianBlur: Don't try to blur an invalid surface
This is a stop-gap measure to deal with invalid surfaces being fed
to this filter.  We get a surface which is not in
CAIRO_STATUS_SUCCESS, and so its width/height are 0.

In reality we should be able to bubble up errors from
cairo_*_surface_create(), but we don't do that yet.