bgo#634514: Don't render unknown elements and their sub-elements
We were representing unknown elements as a NodeGroup, so their
children would be rendered.

Now we represent them as NodeDefs.  This is like a NodeGroup, but it
doesn't do any rendering of itself or its children.  The SVG spec says
that we should stop rendering the file if we encounter an unknown
element, but it seems that real-world renderers are more lenient - so
we'll just not render unknown elements and their children.

https://bugzilla.gnome.org/show_bug.cgi?id=634514