Oops, sort the list of supported SVG features so <switch> works
We use the list of supported features with bsearch(), and the list
was not sorted.

Added struct-cond-03-t.svg from the SVG test suite to this effect.
The reference PNG is different from the official test suite's original
one, because we do NOT support the SVG DOM feature.  Still, the test
file lets us test for <switch> working based on supported feature names.