Backport: build: Check for PangoFT2/FontConfig availability
From commit 46c8688aa1795958fb1308c2bf5ea29295fe6c23

On Windows the GTK+ stack does not hard depend on PangoFT2 (thus
Fontconfig--GTK+ uses PangoWin32 to do the Font discovery and
configuration stuff by default, unless one uses an envvar to force
PangoFT2 usage), unlike *NIX platforms, so we need to check for
it by doing:

-On Windows, enable the test code that uses PangoFT2/FontConfig if
 PangoFT2 and FontConfig is found during configure.  On Visual Studio
 builds, this is set to be disabled in config.h.win32(.in), and can be
 manually enabled by uncommenting #define HAVE_PANGOFT2 1 in
 config.h.win32 prior to the build (or rebuild).  This continues to have
 FontConfig and PangoFT2 to act as an optional dependency.

-On non-Windows platforms, make PangoFT2 and FontConfig a hard dependency,
 which is what the current code assumes.

We might probably need to make the custom TTF load via PangoWin32 and/or
the native Windows API to run the tests when PangoFT2 and FontConfig are
not found on Windows.

Also bump the Pango dependency to 1.38 as the test code uses API that is
introduced in 1.38.x.

https://bugzilla.gnome.org/show_bug.cgi?id=779405