bgo#761175 - Allow masks to (re)use a node being referenced
If a <use> element acquires a node for drawing, then it needs to
release the node before calling pop_discrete_layer().  Mask generation
occurs in that function, and the mask may in turn reuse the referenced
node.  If the node is still acquired at that time, the mask will not
be able to use it.

https://bugzilla.gnome.org/show_bug.cgi?id=761175