Visual Studio projects: Rename PythonPath to PythonDir
This is to avoid confusion with PythonPath, which is commonly used as a
envvar to specify where Python modules are found.

Also split out PythonDir for 32-bit and 64-bit builds, as we are going
to use these for introspection builds directly from the projects, and
set the default Python installations for each Visual Studio version
according to what is now set in GObject-Introspection's Visual Studio
projects.