Visual Studio builds: Integrate introspection in projects
This integrates the introspection build process into the project files,
so that one is able to build the introspection files directly via the
Visual Studio projects.

Note that this is not built by default (and so not cleaned by default),
so one must explicitly select and build the rsvg-introspection project,
which will build the rest of the projects before building the
introspection files.

For this to work, a complete install of GObject-Introspection is
required, with the introspection files for GDK-Pixbuf must be found in
apprpriate locations under $(GlibEtcInstallRoot).