parse_path: Use the absolute value of the rx/ry radii for elliptical arcs
This is per https://www.w3.org/TR/SVG/implnote.html#ArcOutOfRangeParameters

Since we don't need nonnegative_number() anymore, inline it in number().

However, note that
https://www.w3.org/TR/SVG2/paths.html#PathDataErrorHandling indicates
"a non-positive radius value is an error", in SVG2.  At some point we
have to make a decision whether to be permissive, or to actually make
a distinction between parsing SVG1.1 and SVG2.