rsvg-test: Render with a non-identity transformation matrix to test offsetting errors
This is for gitlab#91.  When the initial transformation matrix
in the Cairo context passed to rsvg_handle_render_cairo() is not an
identity matrix (e.g. it has translation offsets), we are not
rendering some things correctly.

So, we now create *all* rendering tests with a nonzero translation.
We do this by creating a "frame" around the destination image, and
then only considering the part inside the frame.