rsvg_parse_style_pair(): Distinguish between styles and presentation attributes
For example, "marker" is a shortcut property that cannot be used from
a presentation attribute:  an SVG element cannot have an attribute
marker="url(#foo)".  This property can only come from the the contents
of the "style" attribute, or from a CSS style.

This makes painting-marker-04-f.svg pass again.