Port PropertyBag to Rust
This is a HashMap for now, to match the C version which used a
GHashTable.  Eventually its implementation will only be a wrapper for
the (char **) from libxml2, plus an iterator over the key/value pairs
in such an array.  For now, we do need lookup() semantics.  We will be
moving the set_atts() methods to iteration instead of lookups one by
one.