Make PDF output reproducible if the SOURCE_DATE_EPOCH envvar is set
Whilst working on the Reproducible Builds effort [0], we noticed
that rsvg-convert does not create reproducible output.

This is because it calls cairo_pdf_surface_create_for_stream
without setting a modification time, thus leading to the current
date always being generated.

An alternative patch *could* use the creation time of the input
file, but this would be misleading metadata when applied to the
output - it was not created at the same time as the input.

This has also been filed in Debian [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/890027

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>