Thoroughly deprecate rsvg_handle_get_metadata().  Returns NULL always.
The code wanted to reconstruct the XML contents of the <metadata>
element so that they could be fetched with the API as a string.
However, it never did this correctly.  So, we return NULL now and
remove the machinery to actually try to reconstruct the XML.

Commit a61da7471f951317e665f90fea51badc57781a2c deprecated this
function, and its comment says:

    Deprecate the title, desc, and metadata accessors

    They're unused except for bindings, according to codesearch. They're also
    implemented incorrectly (only allow retrieving the toplevel's data), and
    did I mention they're completely unused?

Since they were unused back in 2011, we can hopefully safely deprecate
this now.