aspect_ratio: Fix failling test.
parse_align_mode seems like can have 3 states.
* parsed x,y
* parsed None
* Error

Previously I overlooked that and Parsed None was returning the result
as the Error state.
Now the singature is changed to Result<Option<T>, Error> instead of
Option<T> so it will be able to return Ok(None) to indicate the
parsed None state, and Err(E) for the Error State.