Prettify the rather horrible while() loop in resolve_pattern()
The previous commits made it possible to assert that if
pattern.resolve_from_defaults() is called, then a subsequent call to
pattern.is_resolved() will return true.

So, we can now use just this is_resolved() as the "while" loop's
condition, without having to explicitly "break" out of the loop.

This also makes it possible to chain .and_then().and_then().or_else().